Split APU into multiple modules
